agent_app_custom_code/requirements.txt
databricks-sdk 
mlflow 
mlflow-skinny
databricks-vectorsearch 
databricks-agents 
pydantic
pypdf==4.1.0
markdownify==0.12.1
pypandoc_binary==1.13
transformers==4.41.1
torch==2.3.0
tiktoken==0.7.0
langchain-text-splitters==0.2.0
langchain==0.2.11 
langchain_core==0.2.23 
langchain_community==0.2.10